Mempool syncer: organize tests
